requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
google==3.0.0
idna==2.10
lxml==4.6.3
PyMySQL==1.0.2
python-dotenv==0.17.1
python-slugify==5.0.2
requests==2.25.1
soupsieve==2.2.1
text-unidecode==1.3
urllib3==1.26.4
